BENCH: process_global_benchmarks
